+++ Presentation/OpenTgResearcherDesktop/ViewModels/TgClientConnectionViewModel.cs
@@ -66,7 +66,7 @@ public sealed partial class TgClientConnectionViewModel : TgPageViewModelBase
 		IsClientConnected = AppNotificationService.IsClientConnected;
         // Commands
 		ClientConnectCommand = new AsyncRelayCommand(ClientConnectAsync);
-		ClientDisconnectCommand = new AsyncRelayCommand(ClientDisconnectAsync);
+		ClientDisconnectCommand = new AsyncRelayCommand<bool?>(async isQuestion => await ClientDisconnectAsync(isQuestion));
 		AppSaveCommand = new AsyncRelayCommand(AppSaveAsync);
 		AppClearCommand = new AsyncRelayCommand(AppClearAsync);
 		AppDeleteCommand = new AsyncRelayCommand(AppDeleteAsync);
@@ -200,10 +200,35 @@ public sealed partial class TgClientConnectionViewModel : TgPageViewModelBase
 		}
 	}
 
-	private async Task ClientDisconnectAsync() => await ContentDialogAsync(() =>
-        App.BusinessLogicManager.ConnectClient.DisconnectClientAsync(), TgResourceExtensions.AskClientDisconnect(), TgEnumLoadDesktopType.Online);
+	private async Task ClientDisconnectAsync(bool? isQuestion = true) => await ClientDisconnectCoreAsync(isQuestion, isRetry: false);
 
-	private async Task AppLoadCoreAsync()
+    private async Task ClientDisconnectCoreAsync(bool? isQuestion, bool isRetry)
+    {
+        try
+        {
+            Exception.Default();
+            DataRequest = string.Empty;
+            if (isQuestion == true)
+                await ContentDialogAsync(() => App.BusinessLogicManager.ConnectClient.DisconnectClientAsync(), 
+                    TgResourceExtensions.AskClientDisconnect(), TgEnumLoadDesktopType.Online);
+            else
+                await App.BusinessLogicManager.ConnectClient.DisconnectClientAsync();
+        }
+        catch (Exception ex)
+        {
+            LogError(ex);
+            Exception.Set(ex);
+            if (isRetry)
+                return;
+            if (Exception.Message.Contains("or delete the file to start a new session"))
+            {
+                await TgDesktopUtils.DeleteFileAsync(SettingsService.AppSession);
+                await ClientDisconnectCoreAsync(isQuestion, isRetry: true);
+            }
+        }
+    }
+    
+    private async Task AppLoadCoreAsync()
 	{
 		var appResult = await App.BusinessLogicManager.StorageManager.AppRepository.GetCurrentAppAsync(isReadOnly: false);
 		AppEntity = appResult.IsExists && appResult.Item is not null ? appResult.Item : new();